ft_atoi, ft_calloc, ft_memcpy, ft_memmove: Better match libc behaviour
Some of these requirements are questionable, but strict interpretation
of libft assignment requires exactly matching the libc functions'
behaviour. There is allowance for undefined behaviour for crashes, but
it's not clear whether this only applies when libc function would also
crash, and whether this applies even if system manpages don't explicitly
state the behaviour is undefined.