Handle null characters in the file
Because we return C strings, we can't just leave null characters in the
middle of the returned strings - the caller wouldn't know to read past
them, and would lose data between the null character and the next
newline or end of file. Instead, we also break "lines" on null
characters. If a line returned from get_next_line ends without a newline
character, and the next call returns a non-null pointer, the caller can
deduce that the line was broken by a null character. If the file ends
with a null character, get_next_line returns one more empty string to
distinguish from a file ending with any other non-newline character.