Rewrite bonus from scratch to be more readable
By not using separate read_buf, we avoid needing various concatenation
helpers. Instead of appending to line buffer we simply grow it to ensure
it the specified BUFFER_SIZE free space for next read call.

This also changes handling of null characters. Instead of requiring the
caller to get the next line to determine if previous line ended with a
null character or end-of-file, we emit an empty string if and only if
the next character in the file is a null character.